qc/pipeline: move the QC output checking functions to the 'qc/outputs' module.
