+++ auto_process_ngs/test/qc/test_outputs.py
@@ -8,10 +8,13 @@ import shutil
 import os
 from auto_process_ngs.mock import MockIlluminaQcSh
 from auto_process_ngs.mock import MockAnalysisProject
+from auto_process_ngs.mock import UpdateAnalysisProject
 from auto_process_ngs.analysis import AnalysisProject
 from auto_process_ngs.qc.outputs import fastq_screen_output
 from auto_process_ngs.qc.outputs import fastqc_output
 from auto_process_ngs.qc.outputs import fastq_strand_output
+from auto_process_ngs.qc.outputs import check_illumina_qc_outputs
+from auto_process_ngs.qc.outputs import check_fastq_strand_outputs
 from auto_process_ngs.qc.outputs import expected_outputs
 
 # Set to False to keep test output dirs
@@ -62,6 +65,347 @@ class TestFastqStrandOutputFunction(unittest.TestCase):
         self.assertEqual(fastq_strand_output(
             '/data/PB/PB1_ATTAGG_L001_R1_001.fastq.gz'),
                          'PB1_ATTAGG_L001_R1_001_fastq_strand.txt')
+
+class TestCheckIlluminaQcOutputs(unittest.TestCase):
+    """
+    Tests for the 'check_illumina_qc_outputs' function
+    """
+    def setUp(self):
+        # Create a temp working dir
+        self.wd = tempfile.mkdtemp(suffix='TestIlluminaQcOutputs')
+
+    def tearDown(self):
+        # Remove the temporary test directory
+        if REMOVE_TEST_OUTPUTS:
+            shutil.rmtree(self.wd)
+
+    def test_check_illumina_qc_outputs_standardPE_all_missing(self):
+        """
+        check_illumina_qc_outputs: all illumina_qc.sh outputs missing (standardPE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Get the outputs
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="standardPE"),
+                         project.fastqs)
+
+    def test_check_illumina_qc_outputs_standardPE_all_present(self):
+        """
+        check_illumina_qc_outputs: all illumina_qc.sh outputs present (standardPE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Add QC artefacts
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            include_fastq_strand=False,
+            include_multiqc=False)
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="standardPE"),
+                         [])
+
+    def test_check_illumina_qc_outputs_standardPE_some_missing(self):
+        """
+        check_illumina_qc_outputs: some illumina_qc.sh outputs missing (standardPE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Add QC artefacts
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            include_fastq_strand=False,
+            include_multiqc=False)
+        # Remove some outputs
+        for f in ("PJB1_S1_R1_001_fastqc.html",
+                  "PJB1_S1_R1_001_model_organisms_screen.txt",):
+            os.remove(os.path.join(project.qc_dir,f))
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="standardPE"),
+                         [os.path.join(project.fastq_dir,
+                                       "PJB1_S1_R1_001.fastq.gz")])
+
+    def test_check_illumina_qc_outputs_standardSE_all_missing(self):
+        """
+        check_illumina_qc_outputs: all illumina_qc.sh outputs missing (standardSE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Get the outputs
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="standardSE"),
+                         project.fastqs)
+
+    def test_check_illumina_qc_outputs_standardSE_all_present(self):
+        """
+        check_illumina_qc_outputs: all illumina_qc.sh outputs present (standardSE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Add QC artefacts
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            include_fastq_strand=False,
+            include_multiqc=False)
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="standardSE"),
+                         [])
+
+    def test_check_illumina_qc_outputs_standardSE_some_missing(self):
+        """
+        check_illumina_qc_outputs: some illumina_qc.sh outputs missing (standardSE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Add QC artefacts
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            include_fastq_strand=False,
+            include_multiqc=False)
+        # Remove some outputs
+        for f in ("PJB1_S1_R1_001_fastqc.html",
+                  "PJB1_S1_R1_001_model_organisms_screen.txt",):
+            os.remove(os.path.join(project.qc_dir,f))
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="standardSE"),
+                         [os.path.join(project.fastq_dir,
+                                       "PJB1_S1_R1_001.fastq.gz")])
+
+    def test_check_illumina_qc_outputs_singlecell_all_missing(self):
+        """
+        check_illumina_qc_outputs: all illumina_qc.sh outputs missing (singlecell)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Get the outputs
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="singlecell"),
+                         project.fastqs)
+
+    def test_check_illumina_qc_outputs_singlecell_all_present(self):
+        """
+        check_illumina_qc_outputs: all illumina_qc.sh outputs present (singlecell)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Add QC artefacts
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            include_fastq_strand=False,
+            include_multiqc=False)
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="singlecell"),
+                         [])
+
+    def test_check_illumina_qc_outputs_singlecell_some_missing(self):
+        """
+        check_illumina_qc_outputs: some illumina_qc.sh outputs missing (singlecell)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        # Add QC artefacts
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            include_fastq_strand=False,
+            include_multiqc=False)
+        # Remove some outputs
+        for f in ("PJB1_S1_R2_001_fastqc.html",
+                  "PJB1_S1_R2_001_model_organisms_screen.txt",):
+            os.remove(os.path.join(project.qc_dir,f))
+        # Check
+        self.assertEqual(check_illumina_qc_outputs(project,
+                                                   qc_dir="qc",
+                                                   qc_protocol="singlecell"),
+                         [os.path.join(project.fastq_dir,
+                                       "PJB1_S1_R2_001.fastq.gz")])
+
+class TestCheckFastqStrandOutputs(unittest.TestCase):
+    """
+    Tests for the 'fastq_strand_outputs' function
+    """
+    def setUp(self):
+        # Create a temp working dir
+        self.wd = tempfile.mkdtemp(suffix='TestFastqStrandOutputs')
+
+    def tearDown(self):
+        # Remove the temporary test directory
+        if REMOVE_TEST_OUTPUTS:
+            shutil.rmtree(self.wd)
+
+    def test_check_fastq_strand_outputs_standardPE_missing(self):
+        """
+        check_fastq_strand_outputs: fastq_strand.py output missing (standardPE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        # Make fastq_strand.conf
+        fastq_strand_conf = os.path.join(project.dirn,"fastq_strand.conf")
+        with open(fastq_strand_conf,'w') as fp:
+            fp.write("")
+        # Check the outputs
+        self.assertEqual(check_fastq_strand_outputs(project,
+                                                    "qc",
+                                                    fastq_strand_conf,
+                                                    qc_protocol="standardPE"),
+                         [(os.path.join(project.fastq_dir,
+                                        "PJB1_S1_R1_001.fastq.gz"),
+                           os.path.join(project.fastq_dir,
+                                        "PJB1_S1_R2_001.fastq.gz")),])
+
+    def test_check_fastq_strand_outputs_standardPE_present(self):
+        """
+        check_fastq_strand_outputs: fastq_strand.py output present (standardPE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            protocol="standardPE",
+            include_fastq_strand=True,
+            include_multiqc=False)
+        fastq_strand_conf = os.path.join(project.dirn,"fastq_strand.conf")
+        # Check the outputs
+        self.assertEqual(check_fastq_strand_outputs(project,
+                                                    "qc",
+                                                    fastq_strand_conf,
+                                                    qc_protocol="standardPE"),
+                         [])
+
+    def test_check_fastq_strand_outputs_standardSE_missing(self):
+        """
+        check_fastq_strand_outputs: fastq_strand.py output missing (standardSE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        # Make fastq_strand.conf
+        fastq_strand_conf = os.path.join(project.dirn,"fastq_strand.conf")
+        with open(fastq_strand_conf,'w') as fp:
+            fp.write("")
+        # Check the outputs
+        self.assertEqual(check_fastq_strand_outputs(project,
+                                                    "qc",
+                                                    fastq_strand_conf,
+                                                    qc_protocol="standardSE"),
+                         [(os.path.join(project.fastq_dir,
+                                        "PJB1_S1_R1_001.fastq.gz"),),])
+
+    def test_check_fastq_strand_outputs_standardSE_present(self):
+        """
+        check_fastq_strand_outputs: fastq_strand.py output present (standardSE)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            protocol="standardSE",
+            include_fastq_strand=True,
+            include_multiqc=False)
+        fastq_strand_conf = os.path.join(project.dirn,"fastq_strand.conf")
+        # Check the outputs
+        self.assertEqual(check_fastq_strand_outputs(project,
+                                                    "qc",
+                                                    fastq_strand_conf,
+                                                    qc_protocol="standardSE"),
+                         [])
+
+    def test_check_fastq_strand_outputs_singlecell_missing(self):
+        """
+        check_fastq_strand_outputs: fastq_strand.py output missing (singlecell)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        # Make fastq_strand.conf
+        fastq_strand_conf = os.path.join(project.dirn,"fastq_strand.conf")
+        with open(fastq_strand_conf,'w') as fp:
+            fp.write("")
+        # Check the outputs
+        self.assertEqual(check_fastq_strand_outputs(project,
+                                                    "qc",
+                                                    fastq_strand_conf,
+                                                    qc_protocol="singlecell"),
+                         [(os.path.join(project.fastq_dir,
+                                        "PJB1_S1_R2_001.fastq.gz"),),])
+
+    def test_check_fastq_strand_outputs_singlecell_present(self):
+        """
+        check_fastq_strand_outputs: fastq_strand.py output present (singlecell)
+        """
+        # Make mock analysis project
+        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
+                                       "PJB1_S1_R2_001.fastq.gz",),
+                                metadata={ 'Organism': 'Human' })
+        p.create(top_dir=self.wd)
+        project = AnalysisProject("PJB",os.path.join(self.wd,"PJB"))
+        UpdateAnalysisProject(project).add_qc_outputs(
+            protocol="singlecell",
+            include_fastq_strand=True,
+            include_multiqc=False)
+        fastq_strand_conf = os.path.join(project.dirn,"fastq_strand.conf")
+        # Check the outputs
+        self.assertEqual(check_fastq_strand_outputs(project,
+                                                    "qc",
+                                                    fastq_strand_conf,
+                                                    qc_protocol="singlecell"),
+                         [])
 
 class TestExpectedOutputs(unittest.TestCase):
     """
